// peksels/src/entities/enums/filters.ts
export enum SizesFiltersOptions {
  ALL_SIZES = "All Sizes",
  LARGE = "Large",
  MEDIUM = "Medium",
  SMALL = "Small",
}

export enum OrientationsFiltersOptions {
  ALL_ORIENTATIONS = "All Orientations",
  HORIZONTAL = "Horizontal",
  VERTICAL = "Vertical",
  SQUARE = "Square",
}

export enum PhotoSizes {
  LARGE = "5500",
  MEDIUM = "4500",
  SMALL = "3500",
}